refactor(occt-wasm): extract kernel2dOps.ts (~ 800 lines)
Extracts the entire 2D operations section: createPoint2d, createDirection2d,
createVector2d, createAxis2d, wrapCurve2dHandle, makeLine2d, makeCircle2d,
makeArc2dThreePoints, makeArc2dTangent, makeEllipse2d, makeEllipseArc2d,
makeBezier2d, makeBSpline2d, evaluateCurve2d, evaluateCurve2dD1,
getCurve2dBounds, getCurve2dType, trimCurve2d, reverseCurve2d, copyCurve2d,
offsetCurve2d, translateCurve2d, rotateCurve2d, scaleCurve2d,
mirrorCurve2dAtPoint, mirrorCurve2dAcrossAxis, affinityTransform2d,
createIdentityGTrsf2d, createAffinityGTrsf2d, createTranslationGTrsf2d,
createMirrorGTrsf2d, createRotationGTrsf2d, createScaleGTrsf2d,
setGTrsf2dTranslationPart, multiplyGTrsf2d, transformCurve2dGeneral,
intersectCurves2d, projectPointOnCurve2d, distanceBetweenCurves2d,
approximateCurve2dAsBSpline, decomposeBSpline2dToBeziers, createBoundingBox2d,
addCurveToBBox2d, getBBox2dBounds, mergeBBox2d, isBBox2dOut, isBBox2dOutPoint,
getCurve2dCircleData, getCurve2dEllipseData, getCurve2dBezierPoles,
getCurve2dBezierDegree, serializeCurve2d, deserializeCurve2d, splitCurve2d,
liftCurve2dToPlane, buildEdgeOnSurface, extractSurfaceFromFace,
extractCurve2dFromEdge, buildCurves3d, fixWireOnFace.

The c2d/c2dWrap helpers (identity casts at the kernel boundary) move with
the section. ow2d (geometry2d) import moves to kernel2dOps.ts as well.
createCurve2dAdaptor and getCurve2dBSplineData stay as notImplemented stubs
in the adapter, plus fillSurface (also notImplemented).

Adapter: 3058 → 2565 lines (−493). Cumulative reduction from 4341: 41%.
